refactor : extracting utility functions and simplifying tests
