+++ src/__tests__/Minimum1Pizza.test.tsx
@@ -1,8 +1,9 @@
 import { test, expect, beforeAll, vi } from 'vitest';
-import { fireEvent, render, screen } from '@testing-library/react';
+import { fireEvent } from '@testing-library/react';
 
-import App from '../component/App/App';
 import useStore from '../store/store';
+import { OpenModalAndRetrieveButtons } from './utils/OpenModalAndRetrieveButtons';
+import { CheckModalClosure } from './utils/ChekModalClosure';
 
 // Mock la méthode showModal pour éviter les erreurs
 beforeAll(() => {
@@ -10,25 +11,16 @@ beforeAll(() => {
 });
 
 test(`does not allow ordering less than 1 pizza'`, () => {
-  render(<App />);
+  // Ouvre la modal et récupère des boutons (et quantity)
+  const { decrementButton, addToCartButton, quantity } =
+    OpenModalAndRetrieveButtons(1);
 
-  // Récupère le bouton de commande de la première pizza et clique dessus
-  const orderButton = screen.getByTestId('order-button-pizzaId-1');
-  fireEvent.click(orderButton);
-
-  // Récupère et vérifie la présence des boutons 'moins' et 'ajouter au panier'
-  const decrementButton = screen.getByTestId(
-    'modal__quantity-selection-buttons-decrement'
-  );
-  const addToCartButton = screen.getByTestId('modal__add-to-cart-button');
+  // Vérifie la présence du bouton 'moins', qu'il soit bien désactivé, et la présence du bouton 'ajouter au panier'
   expect(decrementButton).toBeInTheDocument();
   expect(decrementButton).toBeDisabled();
   expect(addToCartButton).toBeInTheDocument();
 
   // Clique sur le bouton de décrémentation un nombre alétoire de fois (entre 0 et 50) et vérifie la quantité
-  const quantity = screen.getByTestId(
-    'modal__quantity-selection-buttons-quantity'
-  );
   const randomClicks = Math.floor(Math.random() * 50);
   for (let i = 0; i < randomClicks; i++) {
     fireEvent.click(decrementButton);
@@ -52,9 +44,5 @@ test(`does not allow ordering less than 1 pizza'`, () => {
   expect(shoppingCart).toEqual(expectedShoppingCart);
 
   // Vérifie que la modal est fermée
-  const { quantitySelectionModalisOpen } = useStore.getState();
-  expect(quantitySelectionModalisOpen).toBe(false);
-  expect(screen.queryByTestId('pizza-modal')).not.toBeInTheDocument();
-  expect(decrementButton).not.toBeInTheDocument();
-  expect(addToCartButton).not.toBeInTheDocument();
+  CheckModalClosure();
 });
